perf: lazy load data view charts
